Script to de-embargo data older than a deadline. (#12)
* Script to de-embargo data older than a deadline.

To de-embargo it is to copy it to a local directory and then delete
it from the S3 bucket.

* Finish plumbing, add tqdm progress bars

* black formatting

* Remove unnecessary branch on args.profile

* Drop --continue-on-error

* Individual futures no longer collect errors

* More resumable.

Still isn't behaving as nicely as `aws s3 sync` because ^C will
leave partial files suffixed with a dot and a hex string.  But
it doesn't lose files.

* Respond to PR comments

* Leave the concurrency to S3Transfer.

Doing it on the coarse loop just makes for confusing edge conditions
that don't improve the speed.